requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2.13
django-cors-headers==3.4.0
django-softdelete==0.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
idna==2.10
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
